feat: User now may force usage of this driver
In some cases, the device uses 0xEF00 cluster but doesn't expose it.
It forced me to create an exception list, but it demands an update every time a new device is found in that condition.

So, now, the user may force the usage of this driver and it will recognize as if the device contain the cluster 0xEF00.